+++ actioncable/CHANGELOG.md
@@ -1,2 +1,5 @@
+*   Support large payloads in Action Cable's PostgreSQL adapter.
+
+    *David Backeus*
 
 Please check [7-1-stable](https://github.com/rails/rails/blob/7-1-stable/actioncable/CHANGELOG.md) for previous changes.